feat: Build event handler for clicks, style edit
added event handler for key clicks that recognizes
which key is being pressed and saves the data in
an array.

disabled the ability to click on the container
itself which would result in all key items being
listed out.

removed text cursor from appearing on hover over
keys, but left it for the display so they can
copy data by highlighting it.