feat: Add calculate result function
calcResult function put into selectKey to check
if '=' key is pressed, it is called after cleanInput
to ensure all variables for operate function are
set before running, the calc result is sent to
display but must be put in brackets to ensure the
updateDisplay function receives an array.

firstNumber and secondNumber values needed to be
transformed to parseFloat before being assigned so
operations (+) and decimal results would occur as
expected.