feat: Add chain calculations, restructure code
when an operator symbol is clicked after a result
is presented, it will take the result value as the
new first value and then run a new calculation
against the next value you input along with that
operator.

changed code so display no longer will show the
entire input chain even after calculating results.

if a number key is hit followed by the equals key,
the calculator is cleared back to default.

disabled the ability to click another operator key
until equals or clear is clicked.