fix: Amend last commit and solve multiple decimals
the prior commit only worked for the '+' operator,
a fix to the .includes method now allows the fix to
properly apply to all pertinent operators.

this fix, in turn, appears to have resolved the
multiple decimal point bugs found in a prior
commit, including the inability to add a decimal
in the second segment of an operation after
deleting values up until the operator without
using the '+/-' key two or more times, and the
issue of using the '+/-' key to allow for
additional decimals to be displayed.