feat: Add keyboard support, fix related bug
added keyboard support by creating a new function
with the keyChoice logic from selectKey(), and
then having both event listeners placed inside of
selectKey() with each having access to the new
function when a key is selected.

while testing keyboard input, discovered an issue
when trying to add a leading minus, the default 0
set in cleanInput() when an operator is pressed
would allow multiple operators and ignore the later
half of the equation when calculating the result,
fixed by setting the operator when it is clicked
in those instances.